media-source/SourceBuffer-appendBuffer-null.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>SourceBuffer#appendBuffer(null)</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';

async_test(function(t) {
    var mediaSource = new MediaSource();
    mediaSource.addEventListener('sourceopen', function(e){
        var sourceBuffer = mediaSource.addSourceBuffer(mime);
        t.step(function() {
            assert_throws({name: 'InvalidAccessError'},
                          function() {
                              sourceBuffer.appendBuffer(null);
                          },
                          'SourceBuffer#appendBuffer(null)');
        });
        t.done();
    });
    video.src = window.URL.createObjectURL(mediaSource);
}, 'SourceBuffer#appendBuffer(null) : If data is null then throw an INVALID_ACCESS_ERR exception and abort these steps.');

</script>
</body>
</html>
